Cleaned unecessary fields on the game board, added functionality for 2-4 players, and made the input field more intuitive
