+++ Scrabble/src/View/GameVisualGUI.java
@@ -14,18 +14,16 @@ import java.util.ArrayList;
 
 public class GameVisualGUI extends JFrame {
     private Move previewMove = new Move();
-    private JButton previewButton;
     private JPanel handPanel;
 
     private GameBoard board;
-    private Player p1, p2;
+    private ArrayList<Player> players;
+    private int currentPlayerIndex = 0;
     private PlayerRecords playerSaves;
-    private boolean playerTurn = false;
     private boolean isFirstMove = true;
 
     private JButton[][] boardButtons;
     private JTextField letterField, xField, yField;
-    private JTextArea handArea;
     private JButton submitButton;
 
     private final int BOARD_SIZE = 15; // assuming standard 15x15 board
@@ -74,11 +72,14 @@ public class GameVisualGUI extends JFrame {
 		    final int index = i;
 		    nameFields[i].addActionListener(evt -> {
 		        String name = nameFields[index].getText().trim();
-		        playerSaves.addPlayer(name);
-		        int wins = playerSaves.getWins(name);
-		        int losses = playerSaves.getLosses(name);
-		        recordLabels[index].setText("W/L: " + wins + " / " + losses);
-		      
+		        if (playerSaves.getPlayerNames().contains(name)) {
+			        int wins = playerSaves.getWins(name);
+			        int losses = playerSaves.getLosses(name);
+			        recordLabels[index].setText("W/L: " + wins + " / " + losses);
+		        }
+		        else {
+		        	recordLabels[index].setText("New Player!");;
+		        }
 		
 		        recordLabels[index].revalidate();
 		        recordLabels[index].repaint();
@@ -115,6 +116,7 @@ public class GameVisualGUI extends JFrame {
                 String name = nameFields[i].getText().trim();
                 if (name.isEmpty()) name = "Player " + (i + 1);
                 players.add(new Player(name));
+                playerSaves.addPlayer(name);
             }
             dialog.dispose();
         });
@@ -128,43 +130,24 @@ public class GameVisualGUI extends JFrame {
         dialog.pack();
         dialog.setLocationRelativeTo(this);
         dialog.setVisible(true);
-
+        
+        playerSaves.writePlayerSaves();
         return players;
+        
     }
 
     
     public GameVisualGUI() {
     	playerSaves = new PlayerRecords();
-    	ArrayList<Player> players = showStartupDialog();
+    	players = showStartupDialog();
         board = new GameBoard();
-        p1 = new Player("Johnathan Alexander");
-        p2 = new Player("Erik Picazzo");
 
         setTitle("Scrabble-Style Game - Visual GUI");
         setSize(900, 700);
         setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         setLayout(new BorderLayout());
 
-        // Input panel
-        JPanel inputPanel = new JPanel();
-        letterField = new JTextField(10);
-        xField = new JTextField(3);
-        yField = new JTextField(3);
         submitButton = new JButton("Play Word");
-        previewButton = new JButton("Preview Letter");
-        inputPanel.add(previewButton);
-        previewButton.addActionListener(e -> previewLetter());
-
-
-        inputPanel.add(new JLabel("Letter:"));
-        inputPanel.add(letterField);
-        inputPanel.add(new JLabel("X (1-15):"));
-        inputPanel.add(xField);
-        inputPanel.add(new JLabel("Y (1-15):"));
-        inputPanel.add(yField);
-        inputPanel.add(new JLabel("Direction:"));
-        inputPanel.add(submitButton);
-        add(inputPanel, BorderLayout.NORTH);
 
         // Game board panel (center)
         JPanel boardPanel = new JPanel(new GridLayout(BOARD_SIZE, BOARD_SIZE));
@@ -205,7 +188,7 @@ public class GameVisualGUI extends JFrame {
                             btn.setForeground(Color.BLUE);
 
                             // Optionally store into previewMove
-                            Player current = playerTurn ? p2 : p1;
+                            Player current = players.get(currentPlayerIndex);
                             GamePiece dragged = current.getPiece(letter);
                             if (dragged != null) {
                                 previewMove.addPiece(dragged, finalCol, finalRow);
@@ -230,7 +213,7 @@ public class GameVisualGUI extends JFrame {
         add(boardPanel, BorderLayout.CENTER);
 
         // Player hand panel (bottom)
-        handPanel = new JPanel(new FlowLayout());
+        handPanel = new JPanel(new BorderLayout(10, 0));
         add(handPanel, BorderLayout.SOUTH);
 
         // Action listener
@@ -261,7 +244,7 @@ public class GameVisualGUI extends JFrame {
             return;
         }
 
-        Player current = playerTurn ? p2 : p1;
+        Player current = players.get(currentPlayerIndex);
         GamePiece piece = current.getPiece(letter);
         if (piece == null) {
             JOptionPane.showMessageDialog(this, "You don't have the letter '" + letter + "'");
@@ -297,12 +280,19 @@ public class GameVisualGUI extends JFrame {
         }
 
         handPanel.removeAll();
-        Player current = playerTurn ? p2 : p1;
+        Player current = players.get(currentPlayerIndex);
 
+        // Left: Player name + score
+        JLabel playerInfo = new JLabel(current.getName() + " | Score: " + current.getPlayerPoints());
+        playerInfo.setFont(new Font("SansSerif", Font.BOLD, 16));
+        handPanel.add(playerInfo, BorderLayout.WEST);
+
+        // Center: Player's tiles
+        JPanel tileContainer = new JPanel(new FlowLayout());
         for (GamePiece piece : current) {
             JButton tileButton = new JButton(piece.getLetter());
             tileButton.setFont(new Font("Monospaced", Font.BOLD, 18));
-            tileButton.setTransferHandler(new ValueExportTransferHandler(piece.getLetter())); // custom class below
+            tileButton.setTransferHandler(new ValueExportTransferHandler(piece.getLetter()));
 
             tileButton.addMouseListener(new MouseAdapter() {
                 public void mousePressed(MouseEvent e) {
@@ -312,10 +302,26 @@ public class GameVisualGUI extends JFrame {
                 }
             });
 
-            handPanel.add(tileButton);
+            tileContainer.add(tileButton);
         }
+        handPanel.add(tileContainer, BorderLayout.CENTER);
+
+        // Right: Buttons
+        JPanel buttonPanel = new JPanel(new GridLayout(2, 1, 5, 5));
+
+        JButton removeButton = new JButton("Remove Letters");
+        removeButton.addActionListener(evt -> {
+            previewMove.clear();
+            updateBoardDisplay();
+        });
+
+        buttonPanel.add(submitButton);  // already defined
+        buttonPanel.add(removeButton);
+        handPanel.add(buttonPanel, BorderLayout.EAST);
+
         handPanel.revalidate();
         handPanel.repaint();
+
 
     }
 
@@ -325,7 +331,7 @@ public class GameVisualGUI extends JFrame {
             return;
         }
 
-        Player current = playerTurn ? p2 : p1;
+        Player current = players.get(currentPlayerIndex);
 
         // Create a new Move using the preview
         Move moveToCommit = new Move();
@@ -347,7 +353,7 @@ public class GameVisualGUI extends JFrame {
         isFirstMove = false;
 
         current.fillPlayerHand();
-        playerTurn = !playerTurn;
+        currentPlayerIndex = (currentPlayerIndex + 1) % players.size();
 
         // Clear the preview and update display
         previewMove.clear();
